+++ bundle-analyzer/src/wix-rich-content-plugin-collapsible-list.tsx
@@ -1,5 +1,0 @@
-import 'wix-rich-content-plugin-collapsible-list/dist/styles.min.css';
-import { pluginCollapsibleList } from 'wix-rich-content-plugin-collapsible-list/viewer';
-import { createViewerBundle } from './createViewerBundle';
-
-export default () => createViewerBundle(pluginCollapsibleList());