+++ bundle-analyzer/src/wix-rich-content-plugin-giphy.tsx
@@ -1,5 +1,0 @@
-import 'wix-rich-content-plugin-giphy/dist/styles.min.css';
-import { pluginGiphy } from 'wix-rich-content-plugin-giphy/viewer';
-import { createViewerBundle } from './createViewerBundle';
-
-export default () => createViewerBundle(pluginGiphy());