+++ bundle-analyzer/src/wix-rich-content-plugin-html.tsx
@@ -1,5 +1,0 @@
-import 'wix-rich-content-plugin-html/dist/styles.min.css';
-import { pluginHtml } from 'wix-rich-content-plugin-html/viewer';
-import { createViewerBundle } from './createViewerBundle';
-
-export default () => createViewerBundle(pluginHtml());