+++ bundle-analyzer/src/wix-rich-content-plugin-image.tsx
@@ -1,5 +1,0 @@
-import 'wix-rich-content-plugin-image/dist/styles.min.css';
-import { pluginImage } from 'wix-rich-content-plugin-image/viewer';
-import { createViewerBundle } from './createViewerBundle';
-
-export default () => createViewerBundle(pluginImage());