+++ packages/editor-common/web/src/Utils/inlineStyleUtils.ts
@@ -4,22 +4,25 @@ import {
   isInSelectionRange,
   EditorState,
   Modifier,
-  DraftOffsetKey,
 } from '../index';
 import { uniq } from 'lodash';
 import {
   RICOS_TEXT_HIGHLIGHT_TYPE,
   RICOS_TEXT_COLOR_TYPE,
   RICOS_FONT_SIZE_TYPE,
+  RICOS_BOLD_TYPE,
+  RICOS_ITALIC_TYPE,
   TEXT_COLOR_TYPE,
   TEXT_HIGHLIGHT_TYPE,
   defaultStyleFnMapper,
   CustomInlineStyleType,
-  INLINE_STYLE_TYPES,
   dynamicStyleParsers,
 } from 'wix-rich-content-common';
 
-const getBlockStyleRanges = (block, styleSelectionPredicate?: (style: string) => unknown) => {
+export const getBlockStyleRanges = (
+  block,
+  styleSelectionPredicate?: (style: string) => unknown
+) => {
   const styleRanges: { start: number; end: number; style: string }[] = [];
 
   block.findStyleRanges(
@@ -68,13 +71,10 @@ export const hasOneStyleInSelection = (block, editorState, styleSelectionPredica
   );
 };
 
-/**
- * getSelectionStyles
- *
- * @param {function} styleSelectionPredicate - style selection criteria
- * @returns {string[]} a set of relevant styles found in selection
- */
-export const getSelectionStyles = (styleSelectionPredicate, editorState) => {
+export const getSelectionStyles = (
+  editorState: EditorState,
+  styleSelectionPredicate?: (style: string) => unknown
+) => {
   const selectedBlocks = getSelectedBlocks(editorState);
   return uniq(
     selectedBlocks.reduce((selectedStyles, block) => {
@@ -94,9 +94,12 @@ export const getSelectionStyles = (styleSelectionPredicate, editorState) => {
   );
 };
 
-export const removeCurrentInlineStyle = (editorState: EditorState, styleSelectionPredicate) => {
+export const removeCurrentInlineStyle = (
+  editorState: EditorState,
+  styleSelectionPredicate?: (style: string) => unknown
+) => {
   const selection = editorState.getSelection();
-  const currentStyles = getSelectionStyles(styleSelectionPredicate, editorState);
+  const currentStyles = getSelectionStyles(editorState, styleSelectionPredicate);
   return currentStyles.reduce((nextEditorState, style) => {
     const contentState = nextEditorState.getCurrentContent();
     const nextContentState = Modifier.removeInlineStyle(contentState, selection, style);
@@ -112,12 +115,6 @@ export const setInlineStyle = (editorState: EditorState, inlineStyle: string) =>
     inlineStyle
   );
   return EditorState.push(editorState, contentState, 'change-inline-style');
-};
-
-const typeToCss = {
-  [RICOS_TEXT_COLOR_TYPE]: 'color',
-  [RICOS_TEXT_HIGHLIGHT_TYPE]: 'background-color',
-  [RICOS_FONT_SIZE_TYPE]: 'font-size',
 };
 
 export const getDefaultStyleFn = (type: CustomInlineStyleType) => styles =>
@@ -147,87 +144,26 @@ export const getCustomStyleFns = config => {
     [TEXT_COLOR_TYPE]: colorSetting = {},
     [TEXT_HIGHLIGHT_TYPE]: highlightSettings = {},
   } = config;
+  const colorTypes = [
+    { type: RICOS_TEXT_COLOR_TYPE, setting: colorSetting },
+    { type: RICOS_TEXT_HIGHLIGHT_TYPE, setting: highlightSettings },
+  ];
+  const inlineStyleTypes: CustomInlineStyleType[] = [
+    RICOS_FONT_SIZE_TYPE,
+    RICOS_ITALIC_TYPE,
+    RICOS_BOLD_TYPE,
+  ];
   return [
-    colorSetting.customStyleFn
-      ? getCustomStyleFn(
-          colorSetting.customStyleFn,
-          colorSetting.styleSelectionPredicate,
-          RICOS_TEXT_COLOR_TYPE
-        )
-      : getDefaultStyleFn(RICOS_TEXT_COLOR_TYPE),
-    highlightSettings.customStyleFn
-      ? getCustomStyleFn(
-          highlightSettings.customStyleFn,
-          highlightSettings.styleSelectionPredicate,
-          RICOS_TEXT_HIGHLIGHT_TYPE
-        )
-      : getDefaultStyleFn(RICOS_TEXT_HIGHLIGHT_TYPE),
-    getDefaultStyleFn(RICOS_FONT_SIZE_TYPE),
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    ...colorTypes.map((style: { type: CustomInlineStyleType; setting: any }) => {
+      const {
+        type,
+        setting: { customStyleFn, styleSelectionPredicate },
+      } = style;
+      return customStyleFn
+        ? getCustomStyleFn(customStyleFn, styleSelectionPredicate, type)
+        : getDefaultStyleFn(type);
+    }),
+    ...inlineStyleTypes.map(type => getDefaultStyleFn(type)),
   ];
 };
-
-const getInlineStylesByType = (editorState: EditorState, type: CustomInlineStyleType) => {
-  const styleParser = dynamicStyleParsers[type];
-  return getSelectionStyles(styleParser, editorState).map(style => styleParser(style));
-};
-
-const getSelectionStylesFromDOM = (editorState: EditorState, type: CustomInlineStyleType) => {
-  let currentStyles: (string | null)[] = [];
-  const styleParser = dynamicStyleParsers[type];
-  getSelectedBlocks(editorState)
-    .filter(block => !hasOneStyleInSelection(block, editorState, styleParser))
-    .forEach(block => {
-      const offsetKey = DraftOffsetKey.encode(block.getKey(), 0, 0);
-      const nodes = document.querySelectorAll(`[data-offset-key="${offsetKey}"]`);
-      currentStyles = [
-        ...currentStyles,
-        ...Array.from(nodes).map(node =>
-          window.getComputedStyle(node).getPropertyValue(typeToCss[type])
-        ),
-      ];
-    });
-  return currentStyles;
-};
-
-const setInlineStyleByType = (
-  editorState: EditorState,
-  type: CustomInlineStyleType,
-  style?: string
-) => {
-  const styleParser = dynamicStyleParsers[type];
-  let newEditorState = removeCurrentInlineStyle(editorState, styleParser);
-  if (style) {
-    const inlineStyle = JSON.stringify({
-      [INLINE_STYLE_TYPES[type]]: style,
-    });
-    newEditorState = setInlineStyle(newEditorState, inlineStyle);
-  }
-  return newEditorState;
-};
-
-export const getFontSize = (editorState: EditorState) => {
-  const currentFontSizes = uniq([
-    ...getInlineStylesByType(editorState, RICOS_FONT_SIZE_TYPE),
-    ...getSelectionStylesFromDOM(editorState, RICOS_FONT_SIZE_TYPE),
-  ]);
-  return currentFontSizes.length > 1 || currentFontSizes.length === 0 ? '' : currentFontSizes[0];
-};
-
-export const setFontSize = (editorState: EditorState, data?: { fontSize?: string }) => {
-  let style = data?.fontSize;
-  if (style) {
-    style += 'px';
-  }
-  return setInlineStyleByType(editorState, RICOS_FONT_SIZE_TYPE, style);
-};
-
-export const getColor = (editorState: EditorState, type: CustomInlineStyleType) => {
-  const currentColors = getInlineStylesByType(editorState, type);
-  return currentColors[0];
-};
-
-export const setTextColor = (editorState: EditorState, data?: { color?: string }) =>
-  setInlineStyleByType(editorState, RICOS_TEXT_COLOR_TYPE, data?.color);
-
-export const setHighlightColor = (editorState: EditorState, data?: { color?: string }) =>
-  setInlineStyleByType(editorState, RICOS_TEXT_HIGHLIGHT_TYPE, data?.color);