+++ packages/toolbars-new/web/src/Toolbar/ButtonComponents/NestedMenu.tsx
@@ -21,7 +21,7 @@ type dropDownPropsType = {
 interface NestedMenuProps {
   theme?: RichContentTheme;
   dropDownProps: dropDownPropsType;
-  editorContainer: HTMLElement;
+  getEditorContainer: () => Element;
 }
 
 interface State {
@@ -80,7 +80,7 @@ class NestedMenu extends Component<NestedMenuProps, State> {
   };
 
   render() {
-    const { dropDownProps, theme, editorContainer } = this.props;
+    const { dropDownProps, theme, getEditorContainer } = this.props;
     const { tooltip, dataHook, getIcon, isMobile, t, buttonList } = dropDownProps;
     const { isModalOpen } = this.state;
     return (
@@ -110,7 +110,7 @@ class NestedMenu extends Component<NestedMenuProps, State> {
                 nestedMenu
                 // vertical
                 // afterClick={this.toggleModal}
-                editorContainer={editorContainer}
+                getEditorContainer={getEditorContainer}
               />
             </div>
           )}